Player traversal now works
Before, the player couldn't handle moving from pattern to pattern.
(I forgot to bump order when running off the end of a pattern,
and forgot to loop at the end of the song.)  Fixed now.

I think it's time to decide how to store pitches and noteoff and
blank rows.  And maybe figure out how I'm going to handle
transposes and if they're even a size savings for most songs.